requirements.txt
jax==0.3.15
numpy
protobuf==3.20.2
scikit-learn
tflite==2.4.0
skl2onnx
onnxruntime
onnx_tf
tensorflow-addons[tensorflow]